財務時系列データの桁溢れへの対応

+++ jp.ac.tcu.okadak.ei_mining/src/main/java/jp/ac/tcu/okadak/ei_mining/data_loader/FDBDataLoader.java
@@ -224,8 +224,8 @@ public class FDBDataLoader {
 
 							if (0 != sValue.length()) {
 								// 単位換算する
-								int x = Integer.parseInt(sValue) / UNIT;
-								value = (Double)((double)x);
+								long x = Long.parseLong(sValue) / UNIT;
+								value = (Double) ((double)x);
 							} else {
 								value = null;
 							}
